feat: refactor endpoint routes to clarify when its by course Id and when by subscription Id
