+++ apps/web/src/pages/api/course-subscription/[id]/index.ts
@@ -1,16 +1,12 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { SafeControllerHandling } from '../../../../lib/controllers/SafeControllerHandling';
 import { CoursesApiService } from '../../../../contexts/shared/infrastructure/ApiClients/CoursesApiService';
-import { CourseSubscriptionInfoResponse } from '@studio/commons';
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
   switch (req.method) {
-    case 'GET':
-      await getOwnedCourseSubscriptionByCourseId(req, res);
-      break;
     case 'DELETE':
       await deleteOwnedCourseSubcription(req, res);
       break;
@@ -21,28 +17,6 @@ export default async function handler(
   }
 }
 
-async function getOwnedCourseSubscriptionByCourseId(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  const authToken = req.headers.authorization;
-
-  const coursesApiService = new CoursesApiService();
-
-  const { courseId } = req.query;
-
-  await SafeControllerHandling(res, async () => {
-    const response =
-      await coursesApiService.get<CourseSubscriptionInfoResponse>(
-        `/course-subscription/${courseId}`,
-        undefined,
-        authToken
-      );
-
-    res.status(200).send(response);
-  });
-}
-
 async function deleteOwnedCourseSubcription(
   req: NextApiRequest,
   res: NextApiResponse
@@ -51,7 +25,7 @@ async function deleteOwnedCourseSubcription(
 
   const coursesApiService = new CoursesApiService();
 
-  const { courseId: id } = req.query;
+  const { id } = req.query;
 
   await SafeControllerHandling(res, async () => {
     await coursesApiService.delete<undefined, undefined>(
